Room connection overhaul
Decided it was necessary to distinguish between "clients" connected to a
room via WebSocket, and "players" connected to a game via slots. To wit:
- an unlimited number of clients can connect to a Room
- a Game is instantiated with a finite number of slots

A client starts out as a spectator and can become a player by claiming
an available slot. The first client to connect becomes the manager and
has the option to start the game... which currently does nothing. This
option exists even if the manager is only a spectator.

Built a Room class to manage connections. Need to move it out of api.py.