Add logging, improve slot-release handling
Releasing a slot as the current player now triggers an action that
allows the turn to pass to the next player. If there is no next player,
there will be an error, at least in the pebble game. A poker game could
theoretically proceed - each slot with a stack can pay blinds and fold
even if no client is connected. So handling missing players is a
per-game decision.